requirements.txt
flask==3.1.0
pypdf==3.15.1
python-dotenv==1.1.0
requests==2.32.3
Werkzeug==3.1.3
tiktoken==0.9.0
pydantic==2.11.3